\name{seasondays}
\alias{seasondays}
\docType{data}
\title{
The number of days in a season
}
\description{
The number of days in a season, set to 182.
}

\details{A "season" is taken to be the period over which a model is estimated. In Hyndman & Fan (2010), the season was 1 November to 31 March and so \code{seasondays=151}. The default value in the package is \code{seasondays=182}; i.e., six months. The value of \code{seasondays} can be changed by the user.
}

\keyword{datasets}
