\name{obj.econ}
\alias{obj.econ}
\docType{data}
\title{
Historical demographic & economic data
}
\description{
Seasonal historical demographic & economic data for South Australia
}
\format{
  The format is:
 mts [1:14, 1:6] 1503 1509 1518 1528 1538 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:6] "pop" "gsp" "resiprice" "totalprice" ...
 - attr(*, "tsp")= num [1:3] 2000 2013 1
 - attr(*, "class")= chr [1:2] "mts" "ts"
}
\details{
Seasonal historical demographic & economic data, incluing population, GSP, residential electricity price, total electricity price and cooling/heating degree days 
}
\references{
R. J. Hyndman and S. Fan, "Density Forecasting for Long-term Peak Electricity Demand", IEEE Trans. Power Systems, Vol. 25, No.2, pp. 1142--1153, 2010
}
\examples{
}
\keyword{datasets}
